build: migrate error views
