+++ lib/safira_web/controllers/badge_json.ex
@@ -1,5 +1,4 @@
 defmodule SafiraWeb.BadgeJSON do
-
   alias SafiraWeb.AttendeeJSON
   alias Safira.Avatar
