+++ lib/safira_web/controllers/buy_controller.ex
@@ -48,4 +48,8 @@ defmodule SafiraWeb.BuyController do
         end
     end
   end
+
+  def render(template, _assigns) do
+    %{errors: %{detail: Phoenix.Controller.status_message_from_template(template)}}
+  end
 end
