--- lib/safira_web/views/admin/attendee_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.AttendeeView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
