--- lib/safira_web/views/admin/badge_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.BadgeView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
