--- lib/safira_web/views/admin/company_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.CompanyView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
