--- lib/safira_web/views/admin/redeem_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.RedeemView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
