--- lib/safira_web/views/admin/referral_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.ReferralView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
