--- lib/safira_web/views/admin/staff_view.ex
@@ -1,6 +1,0 @@
-defmodule SafiraWeb.Admin.StaffView do
-  use SafiraWeb, :view
-
-  import Torch.TableView
-  import Torch.FilterView
-end
