--- lib/safira_web/views/error_view.ex
@@ -1,25 +1,0 @@
-defmodule SafiraWeb.ErrorView do
-  use SafiraWeb, :view
-
-  # If you want to customize a particular status code
-  # for a certain format, you may uncomment below.
-  # def render("500.json", _assigns) do
-  #   %{errors: %{detail: "Internal Server Error"}}
-  # end
-
-  # By default, Phoenix returns the status message from
-  # the template name. For example, "404.json" becomes
-  # "Not Found".
-
-  def render("404.json", _assigns) do
-    %{errors: %{detail: "Endpoint Not Found"}}
-  end
-
-  def render("500.json", _assigns) do
-    %{errors: %{detail: "Internal Server Error"}}
-  end
-
-  def template_not_found(template, _assigns) do
-    %{errors: %{detail: Phoenix.Controller.status_message_from_template(template)}}
-  end
-end
