--- lib/safira_web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule SafiraWeb.LayoutView do
-  use SafiraWeb, :view
-end
