--- lib/safira_web/views/pow/registration_view.ex
@@ -1,3 +1,0 @@
-defmodule SafiraWeb.Pow.RegistrationView do
-  use SafiraWeb, :view
-end
