--- lib/safira_web/views/pow/session_view.ex
@@ -1,3 +1,0 @@
-defmodule SafiraWeb.Pow.SessionView do
-  use SafiraWeb, :view
-end
