--- test/safira_web/views/error_view_test.exs
@@ -1,16 +1,0 @@
-defmodule SafiraWeb.ErrorViewTest do
-  use SafiraWeb.ConnCase, async: true
-
-  # Bring render/3 and render_to_string/3 for testing custom views
-  import Phoenix.View
-
-  test "renders 404.json" do
-    assert render(SafiraWeb.ErrorView, "404.json", []) ==
-             %{errors: %{detail: "Endpoint Not Found"}}
-  end
-
-  test "renders 500.json" do
-    assert render(SafiraWeb.ErrorView, "500.json", []) ==
-             %{errors: %{detail: "Internal Server Error"}}
-  end
-end
